backend/requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv
pandas
scikit-learn
flask-cors
